requirements.txt
google-adk==0.1.0
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
Deprecated==1.2.14
